requirements.txt
parallel-web
flask
flask-cors
psycopg2-binary
python-dotenv

